text_recognition/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
cffi==1.15.0
cryptography==37.0.2; python_version >= '3.6'
django-filter==21.1
django-rest-knox==4.2.0
django==4.0.4
djangorestframework==3.13.1
importlib-metadata==4.11.4; python_version < '3.10'
markdown==3.3.7
pillow==9.1.1
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2; python_version >= '3.5'
zipp==3.8.0; python_version >= '3.7'
